% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analogy}
\alias{analogy}
\title{Analogy items dataset}
\format{
A matrix with 149 rows (subjects) and 27 columns (items), where
  each cell contains a binary response (0 = incorrect, 1 = correct).
}
\source{
Blum, D., Holling, H., Galibert, M. S., & Forthmann, B. (2016).
  Task difficulty prediction of figural analogies.
  \emph{Intelligence}, 56, 72--81. \doi{10.1016/j.intell.2016.03.001}
}
\usage{
analogy
}
\description{
Binary item response data from a figural analogies test used to illustrate
the LLTM, MLTM-D, and GMLTM-D models.
}
\references{
Ramirez, E. S., Jimenez, M., Franco, V. R., & Alvarado, J. M. (2024).
Delving into the complexity of analogical reasoning: A detailed exploration
with the Generalized Multicomponent Latent Trait Model for Diagnosis.
\emph{Journal of Intelligence}, \bold{12}, 67.
\doi{10.3390/jintelligence12070067}
}
\keyword{datasets}
