% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability-enhanced.R
\name{bayesian_reliability_fast}
\alias{bayesian_reliability_fast}
\title{Fast Bayesian Marginal Reliability}
\usage{
bayesian_reliability_fast(theta_samples, EAP_theta)
}
\arguments{
\item{theta_samples}{Array of posterior samples [iter, person, component].}

\item{EAP_theta}{Matrix of EAP estimates [person, component].}
}
\value{
List of reliability estimates per component.
}
\description{
Optimized computation of marginal reliability with uncertainty quantification.
}
\keyword{internal}
