% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_reliability.R
\name{conditional_reliability_tif}
\alias{conditional_reliability_tif}
\title{Conditional Reliability based on Test Information Function (TIF)}
\usage{
conditional_reliability_tif(
  fit,
  theta_range = seq(-3, 3, 0.2),
  component = NULL,
  n_samples = 1000
)
}
\arguments{
\item{fit}{Fitted model with \\eqn{\\alpha}, \\eqn{\\beta}, guessing parameters}

\item{theta_range}{Vector of \\eqn{\\theta} values where to evaluate reliability}

\item{component}{Integer or character. Specific component to analyze}

\item{n_samples}{Integer. Number of posterior samples to use}
}
\value{
List with conditional reliability by ability level
}
\description{
Calculates conditional reliability using Test Information Function
for 3-parameter MLTM-D models. This approach is more precise than
quantile-based partitioning methods.
}
