% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability-diagnostics.R
\name{export_reliability_results}
\alias{export_reliability_results}
\title{Export Reliability Results}
\usage{
export_reliability_results(reliability_obj, file_name = NULL)
}
\arguments{
\item{reliability_obj}{Object of class enhanced_mltm_reliability}

\item{file_name}{File name (optional)}
}
\value{
data.frame with tabulated results
}
\description{
Exports results in tabular format for publications.
}
\examples{
\donttest{
  if (!requireNamespace("rstan", quietly = TRUE)) return()
  data(analogy)
  Q <- structure(
    c(0,0,1,0,1,0,1,0,1,1,0,1,1,1,0,1,1,1,0,1,0,1,0,0,1,0,1,
      1,0,0,0,0,1,1,1,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,
      1,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,1,0,1,0,1,1,0,1,0,0,0,
      0,0,0,0,0,0,1,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,0,0,0,0,
      1,0,0,0,1,1,0,1,1,1,1,1,1,1,0,1,1,0,1,1,1,1,0,1),
    dim = c(27L, 5L),
    dimnames = list(NULL, c("rot_fig","rot_trap","reflection",
                            "subt_seg","mov_point")))
  components <- list(global = c(1, 2, 3), local = c(4, 5))
  fit <- GMLTM(analogy, Q, components,
               iters = 200, iter_warmup = 100, chains = 1, cores = 1)
  rel <- enhanced_mltm_reliability(fit, n_samples = 100)
  export_reliability_results(rel)
}

}
\seealso{
Other reliability functions: 
\code{\link{check_reliability_data_quality}()},
\code{\link{demo_reliability_analysis}()},
\code{\link{enhanced_mltm_reliability}()},
\code{\link{quick_reliability_check}()},
\code{\link{reliability}()},
\code{\link{reliability_usage_instructions}()}
}
\concept{reliability functions}
