% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppchecks.R
\name{marginal_Pchecks}
\alias{marginal_Pchecks}
\title{Marginal Proportions Predictive Checks}
\usage{
marginal_Pchecks(fit, interval = 0.95)
}
\arguments{
\item{fit}{MLTM object containing model results.}

\item{interval}{Probability associated with the credibility intervals (default = 0.95).}
}
\value{
A list containing:
\itemize{
 \item \code{items}: A table of fitted values and prediction intervals for each item.
 \item \code{rmsr}: The Root Mean Square Residual (RMSR).
 \item \code{srmr}: The Standardized Root Mean Square Residual (SRMR).
 \item \code{bias}: The difference between the total observed and predicted proportions.
}

The function also generates:
\itemize{
 \item A forest plot visualizing prediction intervals and observed success probabilities.
}
}
\description{
Computes and visualizes marginal success proportions, including predicted values,
confidence intervals, RMSR, SRMR, and bias estimation.
}
\details{
\code{marginal_Pchecks} calculates marginal prediction intervals and observed success proportions.
It prints a table with observed vs. predicted values, generates a forest plot for visualization, and
computes key fit indices: RMSR, SRMR, and bias.
}
