% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability-enhanced.R
\name{plot_comparison_base}
\alias{plot_comparison_base}
\title{Plot Comparisons (Base R)}
\usage{
plot_comparison_base(comparison_data, ...)
}
\description{
Plot Comparisons (Base R)
}
\keyword{internal}
