% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ICC.R
\name{plot_ICC_individual}
\alias{plot_ICC_individual}
\title{Individual Item Characteristic Curves (ICC)}
\usage{
plot_ICC_individual(fit, Q, components)
}
\arguments{
\item{fit}{A fitted GMLTM model object.}

\item{Q}{The Q-matrix for rule-item associations.}

\item{components}{A list indicating rule groupings per component.}
}
\value{
A list of individual \code{ggplot} objects, one per item.
}
\description{
Returns a list of individual ICC \code{ggplot2} plots (one per item).
}
\examples{
\donttest{
  if (!requireNamespace("rstan", quietly = TRUE)) return()
  data(analogy)
  Q <- structure(
    c(0,0,1,0,1,0,1,0,1,1,0,1,1,1,0,1,1,1,0,1,0,1,0,0,1,0,1,
      1,0,0,0,0,1,1,1,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,
      1,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,1,0,1,0,1,1,0,1,0,0,0,
      0,0,0,0,0,0,1,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,0,0,0,0,
      1,0,0,0,1,1,0,1,1,1,1,1,1,1,0,1,1,0,1,1,1,1,0,1),
    dim = c(27L, 5L),
    dimnames = list(NULL, c("rot_fig","rot_trap","reflection",
                            "subt_seg","mov_point")))
  components <- list(global = c(1, 2, 3), local = c(4, 5))
  fit <- GMLTM(analogy, Q, components, iters = 200, iter_warmup = 100, chains = 1)
  plots <- plot_ICC_individual(fit, Q, components)
  print(plots[[1]])
}

}
