% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppchecks.R
\name{ppchecks}
\alias{ppchecks}
\title{Posterior Predictive Checks (PPC) for Model Fit Evaluation}
\usage{
ppchecks(fit, nsim = 100, interval = 0.95, ...)
}
\arguments{
\item{fit}{A fitted MLTM object containing model results.}

\item{nsim}{Number of simulated posterior samples (default = 100).}

\item{interval}{Probability associated with the credibility intervals (default = 0.95).}

\item{...}{Additional graphical parameters to customize the plot.}
}
\value{
A list containing:
\itemize{
  \item \code{items}: A table of fitted values and prediction intervals for each item.
  \item \code{subjects}: Fitted and observed mean scores per subject.
  \item \code{ysim}: Simulated response matrices.
}

The function also generates:
\itemize{
  \item A histogram comparing observed vs. predicted total scores.
}
}
\description{
This function generates posterior predictive checks by plotting the observed and simulated
total scores distribution, comparing empirical data against model predictions. It also
computes fitted values and prediction intervals.
}
\details{
The function simulates multiple datasets from the posterior distribution and compares
the empirical distribution of total scores with the predicted distribution.
It overlays the observed and predicted distributions using a histogram with transparency.

The fitted values, along with their credibility intervals, are computed and returned.
}
\examples{
\donttest{
  if (!requireNamespace("rstan", quietly = TRUE)) return()
  data(analogy)
  Q <- structure(
    c(0,0,1,0,1,0,1,0,1,1,0,1,1,1,0,1,1,1,0,1,0,1,0,0,1,0,1,
      1,0,0,0,0,1,1,1,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,
      1,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,1,0,1,0,1,1,0,1,0,0,0,
      0,0,0,0,0,0,1,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,0,0,0,0,
      1,0,0,0,1,1,0,1,1,1,1,1,1,1,0,1,1,0,1,1,1,1,0,1),
    dim = c(27L, 5L),
    dimnames = list(NULL, c("rot_fig","rot_trap","reflection",
                            "subt_seg","mov_point")))
  fit <- LLTM(analogy, Q, iters = 200, iter_warmup = 100, chains = 1)
  ppchecks(fit)
}

}
\references{
Ramirez, E.S.; Jimenez, M.; Franco, V.R.; Alvarado, J.M. Delving into the Complexity of Analogical Reasoning: A Detailed Exploration with the Generalized Multicomponent Latent Trait Model for Diagnosis. \emph{J. Intell.} 2024, 12, 67. https://doi.org/10.3390/jintelligence12070067
}
