% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLTM.R
\name{print.GMLTM}
\alias{print.GMLTM}
\title{Print method for GMLTM objects}
\usage{
\method{print}{GMLTM}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"GMLTM"}, as returned by \code{\link{GMLTM}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for GMLTM objects
}
