% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLTM_corr.R
\name{print.GMLTM_corr}
\alias{print.GMLTM_corr}
\title{Print method for GMLTM_corr objects}
\usage{
\method{print}{GMLTM_corr}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{c("GMLTM_corr", "GMLTM")}, as returned by
\code{\link{GMLTM_corr}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for GMLTM_corr objects
}
