% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_correlation.R
\name{print.GMLTM_correlation}
\alias{print.GMLTM_correlation}
\title{Print method for GMLTM_correlation objects}
\usage{
\method{print}{GMLTM_correlation}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"GMLTM_correlation"}, as returned by
\code{\link{extract_correlation}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for GMLTM_correlation objects
}
