% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive_check.R
\name{print.GMLTM_prior_predictive_check}
\alias{print.GMLTM_prior_predictive_check}
\title{Print method for GMLTM_prior_predictive_check objects}
\usage{
\method{print}{GMLTM_prior_predictive_check}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"GMLTM_prior_predictive_check"}, as
returned by \code{\link{prior_predictive_check}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for GMLTM_prior_predictive_check objects
}
