% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_report.R
\name{print.GMLTM_student_report}
\alias{print.GMLTM_student_report}
\title{Print Method for GMLTM_student_report Objects}
\usage{
\method{print}{GMLTM_student_report}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"GMLTM_student_report"}, as
returned by \code{\link{student_report}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for GMLTM_student_report Objects
}
