% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLTM.R
\name{print.LLTM}
\alias{print.LLTM}
\title{Print method for LLTM objects}
\usage{
\method{print}{LLTM}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"LLTM"}, as returned by \code{\link{LLTM}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for LLTM objects
}
