% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLTM.R
\name{print.MLTM}
\alias{print.MLTM}
\title{Print method for MLTM objects}
\usage{
\method{print}{MLTM}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"MLTM"}, as returned by \code{\link{MLTM}}.}

\item{digits}{Number of decimal digits to display. Default is 3.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for MLTM objects
}
