% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{print.reliability_data_quality}
\alias{print.reliability_data_quality}
\title{Print Method for Data Quality Diagnostics}
\usage{
\method{print}{reliability_data_quality}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"reliability_data_quality"}.}

\item{...}{Currently unused.}
}
\description{
Print Method for Data Quality Diagnostics
}
