% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_reliability.R
\name{reliability_profile}
\alias{reliability_profile}
\title{Quick Reliability Profile Analysis}
\usage{
reliability_profile(cond_rel_obj)
}
\arguments{
\item{cond_rel_obj}{An object of class \code{conditional_reliability_tif} returned by
\code{\link{conditional_reliability_tif}}.}
}
\value{
A list of class \code{"reliability_profile"} with summary metrics per component.
}
\description{
Provides a quick overview of reliability characteristics for each component.
}
