% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_reliability.R
\name{compare_conditional_reliability}
\alias{compare_conditional_reliability}
\title{Compare Conditional Reliability Between Components at Specific Theta Values}
\usage{
compare_conditional_reliability(cond_rel_obj, theta_points = c(-1, 0, 1))
}
\arguments{
\item{cond_rel_obj}{An object of class \code{conditional_reliability_tif} returned by
\code{\link{conditional_reliability_tif}}.}

\item{theta_points}{Numeric vector of theta values at which to compare components.
Default is \code{c(-1, 0, 1)}.}
}
\description{
Compare Conditional Reliability Between Components at Specific Theta Values
}
