% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{compute_model_validation}
\alias{compute_model_validation}
\title{Compute LOO and WAIC for GMLTM models}
\usage{
compute_model_validation(fit)
}
\arguments{
\item{fit}{A fitted GMLTM model or a list of fitted models.}
}
\value{
If a single model is provided, returns a list with LOO and WAIC results.
If multiple models are provided, returns a summary table with key LOO and WAIC
indices.
}
\description{
This function extracts the log-likelihood from a GMLTM model
and computes the Leave-One-Out Cross-Validation (LOO) and
the Widely Applicable Information Criterion (WAIC).
LOO is a Bayesian model comparison metric based on Pareto-smoothed importance
sampling, while WAIC is a fully Bayesian criterion that estimates predictive
accuracy.
}
\examples{
\donttest{
  if (!requireNamespace("rstan", quietly = TRUE)) return()
  data(analogy)
  Q <- structure(
    c(0,0,1,0,1,0,1,0,1,1,0,1,1,1,0,1,1,1,0,1,0,1,0,0,1,0,1,
      1,0,0,0,0,1,1,1,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,
      1,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,1,0,1,0,1,1,0,1,0,0,0,
      0,0,0,0,0,0,1,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,0,0,0,0,
      1,0,0,0,1,1,0,1,1,1,1,1,1,1,0,1,1,0,1,1,1,1,0,1),
    dim = c(27L, 5L),
    dimnames = list(NULL, c("rot_fig","rot_trap","reflection",
                            "subt_seg","mov_point")))
  components <- list(global = c(1, 2, 3), local = c(4, 5))
  fit1 <- GMLTM(data = analogy, Q = Q, components = components,
                iters = 200, iter_warmup = 100, chains = 1)
  compute_model_validation(fit1)
}
}
\references{
Vehtari, A., Gelman, A., & Gabry, J. (2017). Practical Bayesian model
evaluation using LOO-CV and WAIC. Statistics and Computing, 27(5), 1413--1432.
\doi{10.1007/s11222-016-9696-4}
}
