% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{export_reliability_results}
\alias{export_reliability_results}
\title{Export Reliability Results}
\usage{
export_reliability_results(reliability_obj, file_name = NULL)
}
\arguments{
\item{reliability_obj}{Object of class enhanced_mltm_reliability}

\item{file_name}{File name (optional)}
}
\value{
data.frame with tabulated results
}
\description{
Exports results in tabular format for publications.
}
