% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppchecks.R
\name{marginal_Pchecks}
\alias{marginal_Pchecks}
\title{Marginal Proportions Predictive Checks}
\usage{
marginal_Pchecks(fit, interval = 0.95)
}
\arguments{
\item{fit}{MLTM object containing model results.}

\item{interval}{Probability associated with the credibility intervals (default = 0.95).}
}
\value{
A list containing:
\itemize{
 \item \code{items}: A table of fitted values and prediction intervals for each item.
 \item \code{rmsr}: The Root Mean Square Residual (RMSR).
 \item \code{srmr}: The Standardized Root Mean Square Residual (SRMR).
 \item \code{bias}: The difference between the total observed and predicted proportions.
}

The function also generates:
\itemize{
 \item A forest plot visualizing prediction intervals and observed success probabilities.
}
}
\description{
Computes and visualizes marginal success proportions, including predicted values,
confidence intervals, RMSR, SRMR, and bias estimation.
}
\details{
\code{marginal_Pchecks} calculates marginal prediction intervals and observed success proportions.
It prints a table with observed vs. predicted values, generates a forest plot for visualization, and
computes key fit indices: RMSR, SRMR, and bias.
}
\examples{
\donttest{
  if (!requireNamespace("rstan", quietly = TRUE)) return()
  data(analogy)
  Q <- structure(
    c(0,0,1,0,1,0,1,0,1,1,0,1,1,1,0,1,1,1,0,1,0,1,0,0,1,0,1,
      1,0,0,0,0,1,1,1,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,
      1,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,1,0,1,0,1,1,0,1,0,0,0,
      0,0,0,0,0,0,1,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,0,0,0,0,0,
      1,0,0,0,1,1,0,1,1,1,1,1,1,1,0,1,1,0,1,1,1,1,0,1),
    dim = c(27L, 5L),
    dimnames = list(NULL, c("rot_fig","rot_trap","reflection",
                            "subt_seg","mov_point")))
  fit <- LLTM(analogy, Q, iters = 200, iter_warmup = 100, chains = 1, cores = 1)
  marginal_Pchecks(fit)
}

}
\references{
Ramirez, E. S., Jimenez, M., Franco, V. R., & Alvarado, J. M. (2024).
Delving into the complexity of analogical reasoning: A detailed exploration
with the Generalized Multicomponent Latent Trait Model for Diagnosis.
\emph{Journal of Intelligence}, \bold{12}, 67.
\doi{10.3390/jintelligence12070067}
}
\seealso{
Other posterior predictive check functions: 
\code{\link{ppchecks}()}
}
\concept{posterior predictive check functions}
