% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability-enhanced.R
\name{plot_marginal_base}
\alias{plot_marginal_base}
\title{Plot Marginal Reliability (Base R)}
\usage{
plot_marginal_base(marginal_data, ...)
}
\description{
Plot Marginal Reliability (Base R)
}
\keyword{internal}
