% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{print.enhanced_mltm_reliability}
\alias{print.enhanced_mltm_reliability}
\title{Print Method for Enhanced MLTM Reliability}
\usage{
\method{print}{enhanced_mltm_reliability}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"enhanced_mltm_reliability"}.}

\item{digits}{Integer. Number of decimal places to display. Default is \code{3}.}

\item{...}{Currently unused.}
}
\description{
Print Method for Enhanced MLTM Reliability
}
